src/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "The environment server that makes it possible to interact with the environment."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "numpy>=2.2.4",
    "pillow>=11.1.0",
    "pyautogui>=0.9.54",
    "uvicorn>=0.34.2",
    # "websockets>=15.0",
]
